+++ packages/iov-bns/src/index.ts
@@ -15,7 +15,6 @@ export {
   // transactions
   BnsTx,
   AddAddressToUsernameTx,
-  SetNameTx,
   RegisterBlockchainTx,
   RegisterUsernameTx,
   RemoveAddressFromUsernameTx,