+++ packages/iov-cli/src/cli.ts
@@ -45,7 +45,6 @@ export function main(originalArgs: ReadonlyArray<string>): void {
   console.log(colors.yellow("  * from @iov/bns"));
   console.log(colors.yellow("    - bnsCodec"));
   console.log(colors.yellow("    - bnsConnector"));
-  console.log(colors.yellow("    - SetNameTx"));
   console.log(colors.yellow("  * from @iov/core"));
   console.log(colors.yellow("    - Address"));
   console.log(colors.yellow("    - ChainId"));
@@ -88,7 +87,6 @@ export function main(originalArgs: ReadonlyArray<string>): void {
     import {
       bnsCodec,
       bnsConnector,
-      SetNameTx,
     } from '@iov/bns';
     import {
       Address,